+++ src/lib/BuildError.ts
@@ -1,0 +1,5 @@
+// Exports
+export default class BuildError extends Error {
+  public errors: string[] = [];
+  public warnings: string[] = [];
+}