+++ src/lib/getConfigs.ts
@@ -1,5 +1,12 @@
+// Dependencies
+import webpack from 'webpack';
+
+// Types
+export type ValidConfiguration<T = any> = string | webpack.Configuration | ((opts: T) => webpack.Configuration);
+export type ValidConfigurations<T = any> = ValidConfiguration<T> | Array<ValidConfiguration<T>>;
+
 // Exports
-module.exports = (configFilesOrObjectsOrFns, opts) => {
+export default function getConfigs<T>(configFilesOrObjectsOrFns: ValidConfigurations<T>, opts?: T) {
   const configs = (Array.isArray(configFilesOrObjectsOrFns)
     ? configFilesOrObjectsOrFns
     : [configFilesOrObjectsOrFns]
@@ -13,4 +20,4 @@ module.exports = (configFilesOrObjectsOrFns, opts) => {
     const rawConfig = typeof config === 'string' ? require(config) : config;
     return typeof rawConfig === 'function' ? rawConfig(opts) : rawConfig;
   });
-};
+}