+++ src/lib/logger.js
@@ -1,1 +1,0 @@
-module.exports = (opts = {}) => (output = '') => !opts.silent && console.log(output.toString());