+++ test/__tests__/lib/getCompilerStats.ts
@@ -1,5 +1,5 @@
 // Dependencies
-const getCompilationStats = require('../../../src/lib/getCompilationStats');
+import getCompilerStats from '../../../src/lib/getCompilerStats';
 
 // Tests
 describe('lib', () => {
@@ -15,12 +15,12 @@ describe('lib', () => {
           endTime: 1531090474688,
         },
       ];
-      const buildStats = getCompilationStats(mockStats);
-      expect(buildStats.timings.total.duration).toBe(3785);
-      expect(buildStats.timings.total.minStart).toBe(1531090470903);
-      expect(buildStats.timings.total.minEnd).toBe(1531090472903);
-      expect(buildStats.timings.total.maxStart).toBe(1531090471183);
-      expect(buildStats.timings.total.maxEnd).toBe(1531090474688);
+      const buildStats = getCompilerStats(mockStats as any);
+      expect(buildStats.timings.duration).toBe(3785);
+      expect(buildStats.timings.minStart).toBe(1531090470903);
+      expect(buildStats.timings.minEnd).toBe(1531090472903);
+      expect(buildStats.timings.maxStart).toBe(1531090471183);
+      expect(buildStats.timings.maxEnd).toBe(1531090474688);
       expect(buildStats.timings.perCompiler[0].start).toBe(1531090470903);
       expect(buildStats.timings.perCompiler[0].end).toBe(1531090472903);
       expect(buildStats.timings.perCompiler[0].duration).toBe(2000);
@@ -34,12 +34,12 @@ describe('lib', () => {
         startTime: 1531090471182,
         endTime: 1531090474688,
       };
-      const buildStats = getCompilationStats(mockStats);
-      expect(buildStats.timings.total.duration).toBe(3506);
-      expect(buildStats.timings.total.minStart).toBe(1531090471182);
-      expect(buildStats.timings.total.minEnd).toBe(1531090474688);
-      expect(buildStats.timings.total.maxStart).toBe(1531090471182);
-      expect(buildStats.timings.total.maxEnd).toBe(1531090474688);
+      const buildStats = getCompilerStats(mockStats as any);
+      expect(buildStats.timings.duration).toBe(3506);
+      expect(buildStats.timings.minStart).toBe(1531090471182);
+      expect(buildStats.timings.minEnd).toBe(1531090474688);
+      expect(buildStats.timings.maxStart).toBe(1531090471182);
+      expect(buildStats.timings.maxEnd).toBe(1531090474688);
       expect(buildStats.timings.perCompiler[0].start).toBe(1531090471182);
       expect(buildStats.timings.perCompiler[0].end).toBe(1531090474688);
       expect(buildStats.timings.perCompiler[0].duration).toBe(3506);